Develop (#4)
* fix: ensure recentPassRate always returns an integer

* fix: add WebKit-specific styles for cross-browser flip card compatibility

* feat: add question review with filtering and display breakdown in test results

* fix: remove null-safe operator where values are always present in TestResult

+++ app/Models/TestSession.php
@@ -35,7 +35,7 @@ class TestSession extends Model
 
     public static function recentPassRate(int $days = 30): int
     {
-        return Cache::remember("pass_rate_{$days}", 600, function () use ($days) {
+        return (int) Cache::remember("pass_rate_{$days}", 600, function () use ($days) {
             $cutoff = now()->subDays($days);
 
             $total = static::completed()
